tflite-code/requirements.txt
numpy
Pillow
tflite-runtime
